HANDLEを扱う部分のエラー対策
git-svn-id: https://svn.osdn.net/svnroot/ttssh2/branches/x64@7870 f5f01b69-1e22-0410-acbf-894ab4bd6246

+++ TTProxy/ProxyWSockHook.h
@@ -618,12 +618,14 @@ private:
         HANDLE getTask(ConnectionInfo* info) {
             if (info == NULL)
                 return NULL;
-            return (HANDLE) -info->addr.S_un.S_un_b.s_b4;
+			// s_b4のマイナスの値をハンドルとして使用する
+            return (HANDLE)(intptr_t)-info->addr.S_un.S_un_b.s_b4;
         }
         ConnectionInfo* get(HANDLE task) {
-            if ((DWORD) task >= 0)
+			// ハンドルは数値としてマイナスの値のはず
+            if ((intptr_t) task >= 0)
                 return NULL;
-            return get((int) -((long) task) - 1);
+            return get((int) -((intptr_t) task) - 1);
         }
         ConnectionInfo* get(in_addr addr) {
             if (addr.S_un.S_un_b.s_b1 != 0 || addr.S_un.S_un_b.s_b2 != 0 || addr.S_un.S_un_b.s_b3 != 0)
